documentation
add lpc178x user manual

+++ application/inc/display.h
@@ -1,7 +1,7 @@
 #ifndef __DISPLAY_H_
 #define __DISPLAY_H_
 
-extern uint32_t display_fb[800][480];
+extern uint16_t display_fb[800][480];
 extern void	Initialize_Display();
